+++ src/ConfigMenu.cs
@@ -20,6 +20,7 @@ internal static class GGVConfig
   internal static bool FIX_HIGH_STRESS      = true;
   internal static bool FIX_BRICK_OF_CASH    = true;
   internal static bool FIX_GUN_MAGNIFICENCE = true;
+  internal static bool FIX_INFINITE_BEAMS   = true;
 
   // Safe Optimizations
   internal static int  PREALLOCATE_HEAP  = 0;
@@ -80,6 +81,7 @@ internal static class GGVConfig
     FIX_HIGH_STRESS      = "Enabled" == ConfigMenu._Gunfig.Value(ConfigMenu.HIGH_STRESS);
     FIX_BRICK_OF_CASH    = "Enabled" == ConfigMenu._Gunfig.Value(ConfigMenu.BRICK_OF_CASH);
     FIX_GUN_MAGNIFICENCE = "Enabled" == ConfigMenu._Gunfig.Value(ConfigMenu.GUN_MAGNIFICENCE);
+    FIX_INFINITE_BEAMS   = "Enabled" == ConfigMenu._Gunfig.Value(ConfigMenu.INFINITE_BEAMS);
 
     OPT_VIS_CHECKS       = "Enabled" == ConfigMenu._Gunfig.Value(ConfigMenu.VIS_CHECKS);
     OPT_OCCLUSION        = "Enabled" == ConfigMenu._Gunfig.Value(ConfigMenu.OCCLUSION);
@@ -127,6 +129,7 @@ internal static class GGVConfig
     WriteLine($"FIX_HIGH_STRESS          = {FIX_HIGH_STRESS}");
     WriteLine($"FIX_BRICK_OF_CASH        = {FIX_BRICK_OF_CASH}");
     WriteLine($"FIX_GUN_MAGNIFICENCE     = {FIX_GUN_MAGNIFICENCE}");
+    WriteLine($"FIX_INFINITE_BEAMS       = {FIX_INFINITE_BEAMS}");
 
     WriteLine($"OPT_OCCLUSION            = {OPT_OCCLUSION}");
     WriteLine($"OPT_AMMO_DISPLAY         = {OPT_AMMO_DISPLAY}");
@@ -191,6 +194,7 @@ internal static class ConfigMenu
   internal const string HIGH_STRESS      = "High Stress Armor Fix";
   internal const string BRICK_OF_CASH    = "Brick of Cash Fix";
   internal const string GUN_MAGNIFICENCE = "Gun Magnificence Fix";
+  internal const string INFINITE_BEAMS   = "Infinite Beam Ammo Fix";
 
   internal const string SAFE_OPT         = "Safe Optimizations";
   internal const string PREALLOCATE      = "Preallocate Heap Memory";
@@ -247,6 +251,7 @@ internal static class ConfigMenu
     sf.FancyToggle(HIGH_STRESS, "Fixes no health or armor being displayed\nduring High Stress mode for Robot or\nother 0-health characters.");
     sf.FancyToggle(BRICK_OF_CASH, "Fixes secret room Snitch spawned by Brick of Cash\nitem from spawning inside walls on\nthe right side of rooms.");
     sf.FancyToggle(GUN_MAGNIFICENCE, "Fixes magnificence increasing when\npicking up a gun that has already been picked up.");
+    sf.FancyToggle(INFINITE_BEAMS, "Fixes beams depleting ammo even when under\nthe effects of Magazine Rack.");
 
     Gunfig so = _Gunfig.AddSubMenu(SAFE_OPT);
     so.FancyMemList(PREALLOCATE, "Preallocates RAM to avoid OS requests later.\nDefault uses Gungeon's default of about 200MB.\nHigher values result in fewer lag spikes.");